Create a for loop in order to read all the data for student `S1`.
In this case, first I create a list of dictionaries called `physio_variables` in which I'm saving the name of the variable
based on the original file name and the name of the column or columns based of info.txt and readme.

I also create a list for each of the three exams and a list for each student, following original filename.

After that, I generate a loop in which I iterate over each type of exam, then over each variable, and then over the
dictionary that saves the columns information for each variable.

I use f-strings to save information of each iteration, reading each csv file with pd.read_csv to use DataFrames. And I
append each DataFrame to a list.

TODO: Make the iterations over each student in order to upload all the information just from the same cell.