Add a dictionary for registering the initial time and the sample rate of each student variable.
Generate a new function in order to extract the physiological data for every participant. The function reads information
related with the physiological variable, students and exam, and generate a dictionary using strings and numpy arrays in
order to save data from the same sensor from different students.

Make a first approach to data visualization in order to check the status of the recordings and eliminate in the future
the neccesary outliers.