Generate polynomial fit in order to detrend the Heart Rate signal for student 10 in midterm 2 exam.
The order calculated using Bayes information criterion is 32. Then, the filter is generated based on the residual of
the original signal and  the estimated coefficients in y_hat. In order to correct offset, the absolute value of the minimum of the signal is added to the filtered signal.

Both, Bayes information criterion and filtered signal are plotted as a subplot.

Also, the time vector (in seconds) and frequency vector (in Hz) are generated. After that, FFT of filtered signal is
calculated and the magnitude of FFT is plotted using stem.