Add new variables in order to generata different plots based on student, exam and variable.
Identify that each student has an specific start point and cut the plot edges in which the experiment (exam)
has not started. Also, extract unix time and convert to datetime, using `utcfromtimestamp` fuction.

Draw inferior and superior limits (boundaries) for each variable with vertical intermitent lines, locating them
in the points in which the plot is aproximmately 3000 and 10000. However, those values are arbitrary based on some
student, so it is necessary to adjust them based on local maximums and minimums.

Finally, generate histogram of the cutted and filtered version of the signal.

It is important to take into account that the detrending filter might not be neccesary because the oscillations are
apparently mainly generated due to the start and finish of the exam, and those boundaries were already skipped.