Add titles and subtitles to all the project in order to divide the methodology stages. Most of the code was split in different cells for better organization. Also, the first main description of the project was modified.
Modified variable `physio_variables` in order to leave it only with the neccesary variables. In this case, I work only
with EDA, BVP and TEMP. The variable neccesary to read all the information know is `default_variables`.

Comment zip_extraction function, due to is not neccesary to execut them anymore.

Rename the students grade dataset columns for convenience.

Generate a function in order to visualize simultaneously all the data of the students based on the specified variables.
The function also has the possibility of showing a barplot in order to compare the performance of the student versus
the mean, and observe the variability on physiological signals.