Identify boundaries for cutting the edges of the signals based on the peaks on EDA signal.
Define threshold based on percentile 80 of the signal and identify all the data points over threshold in order to use the first and the last one as the boundaries for the signal.

Generate a for loop plotting in order to visualize the efectiveness of boundaries, for different students and exam signals.