Generate a function `cut_boundaries` in order to indentify the location of the lower and upper bound for each of the signals based on the data obtained from EDA.
The function detect the range in which the physiological data was correctly acquired.

It is necessary to configure the function in order to return an array with the new information. Also, the function code can be might be generalized for any selected signal, so it is needed to adapt the function.