Adapt new dataset in order to have a full structure taking into account the exam, the statistical measure and the physiological variable in each of the columns in order to have a final dataset with 10 rows (students) and 30 colums.
Compute the modified z-score outlier removal method for each of the signals recorded and equal all the outlier values to np.nan. The outlier threshold is 2.5. After identifying outliers, global maximum, median and variance of each preprocessed signal, are calculated in order to make statistical analyses posteriorly.

Generate a subplot in order to compare the different signals after outlier removal.

Modify size of each plot in the general visualization of the dataset, for visualization comfort.